+++ apps/app/app/(demo)/demo/@send/default.tsx
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import OriginalDefault from "@/app/(wallet)/[address]/@send/default";
+import OriginalDefault from "@/app/(wallet)/@send/default";
 
 // -----------------------------------------------------------------------------
 // Original Default
