+++ apps/app/app/(wallet)/@notifications/(.)notifications/page.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable jsx-a11y/img-redundant-alt */
-/* eslint-disable @next/next/no-img-element */
 // Copyright (C) 2023 Light, Inc.
 //
 // This program is free software: you can redistribute it and/or modify
@@ -15,26 +13,31 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import OriginalPage from "@/app/(wallet)/@op/(.)[address]/op/page";
-import { DEMO_WALLET_ADDRESS } from "@/const";
+import OriginalPage from "@/app/(authenticated)/notifications/page";
+import { Modal } from "@/components/modal";
 
 // -----------------------------------------------------------------------------
 // Props
 // -----------------------------------------------------------------------------
 
-interface PageProps {
+type PageProps = {
   searchParams: {
-    userOperations?: string;
+    pagination?: string;
   };
-}
+};
 
 // -----------------------------------------------------------------------------
-// Original Page
+// Page
 // -----------------------------------------------------------------------------
 
 export default async function Page({ searchParams }: PageProps) {
-  return OriginalPage({
-    params: { address: DEMO_WALLET_ADDRESS },
-    searchParams,
-  });
+  // ---------------------------------------------------------------------------
+  // Render
+  // ---------------------------------------------------------------------------
+
+  return (
+    <Modal>
+      <OriginalPage searchParams={searchParams} />
+    </Modal>
+  );
 }
