+++ apps/app/app/(wallet)/error.tsx
